requirements.txt
functions-framework==3.2.0
google-cloud-storage==2.7.0
